Add coverage of privileged tests
