+++ pkg/ebpf/common/common_privileged_test.go
@@ -1,6 +1,8 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
+//go:build privileged_tests
+
 package ebpfcommon
 
 import (
@@ -12,12 +14,12 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func setIntegrity(t *testing.T, path, text string) {
-	err := os.WriteFile(path, []byte(text), 0o644)
+func setNotReadable(t *testing.T, path string) {
+	err := os.Chmod(path, 0o00)
 	require.NoError(t, err)
 }
 
-func TestLockdownParsing(t *testing.T) {
+func TestLockdownParsing_Privileged(t *testing.T) {
 	noFile, err := os.CreateTemp(t.TempDir(), "not_existent_fake_lockdown")
 	require.NoError(t, err)
 	notPath, err := filepath.Abs(noFile.Name())
@@ -39,18 +41,7 @@ func TestLockdownParsing(t *testing.T) {
 	// Setup for testing
 	lockdownPath = path
 
-	setIntegrity(t, path, "none [integrity] confidentiality\n")
-	assert.Equal(t, KernelLockdownIntegrity, KernelLockdownMode())
-
 	setIntegrity(t, path, "[none] integrity confidentiality\n")
-	assert.Equal(t, KernelLockdownNone, KernelLockdownMode())
-
-	setIntegrity(t, path, "none integrity [confidentiality]\n")
-	assert.Equal(t, KernelLockdownConfidentiality, KernelLockdownMode())
-
-	setIntegrity(t, path, "whatever\n")
-	assert.Equal(t, KernelLockdownOther, KernelLockdownMode())
-
-	setIntegrity(t, path, "")
+	setNotReadable(t, path)
 	assert.Equal(t, KernelLockdownIntegrity, KernelLockdownMode())
 }
