+++ pkg/internal/ebpf/verifier/bpf_verifier_test.go
@@ -1,7 +1,7 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-//go:build linux
+//go:build linux && bpf_verifier_tests
 
 package bpf_verifier_test
 
@@ -30,8 +30,6 @@ import (
 	rdnsxdpbpf "go.opentelemetry.io/obi/pkg/internal/rdns/ebpf/xdp"
 	statsolly "go.opentelemetry.io/obi/pkg/internal/statsolly/ebpf"
 )
-
-const privilegedEnv = "PRIVILEGED_TESTS"
 
 // loadAndVerify loads a BPF collection spec into the kernel, triggering the BPF
 // verifier, then immediately closes it. Any verifier rejection surfaces as a test failure.
@@ -137,10 +135,6 @@ func forEachCombination(t *testing.T, prefix string, loadFn func() (*ebpf.Collec
 // Requires CAP_SYS_ADMIN / root.
 // Run with: sudo env PATH=$PATH PRIVILEGED_TESTS=true go test ./pkg/internal/ebpf/verifier/...
 func TestBPFVerifierWithConstants(t *testing.T) {
-	if os.Getenv(privilegedEnv) == "" {
-		t.Skipf("Skipping this test because %v is not set", privilegedEnv)
-	}
-
 	if err := rlimit.RemoveMemlock(); err != nil {
 		t.Skipf("cannot remove memlock limit (insufficient privileges?): %v", err)
 	}
